Moves run_emails() to web_scraping.py
Run_emails() was a defined function inside fetch_and_alert.py,
but it was not being called upon as it was for a previous
iteration of fetch_and_alert.py. It's utility is for an old
camera, bigfoot, that sent every photo to an email, and
run_emails() had to scrape the email it was offloading into,
grab the new images, and then process them as normal
in fetch_and_alert.py.

Run_emails is now in web_scraping, where web_scraping was
adjusted to include the libraries from fetch_emails.py that
run_emails() calls upon, and fetch_and_alert.py was adjusted
to no longer include the libraries that run_emails()
depends on and also no longer includes run_emails().