Move detect function into separate file
fetch_and_alert.py has a lot of clutter with function
definitions and old comments and dependencies that are no
longer used. The function that uses megadetector and
puts bounding boxes on the images was a function defined
within fetch_and_alert and then called later, but it makes
more sense for it to be separate in case this function needs
to be used elsewhere.