Reconfigure detect_img so that is does not read in the config file
detect_img was taking the config file from fetch_and_alert and
reading it and taking the variables it needed each time, but
this was redundant as fetch_and_alert was already parsing the
config file. Now detect_img uses the global variables obtained
in fetch_and_alert and does not parse the .yml itself.

There were also missing variable declarations within detect_img
that are now in there.