Store recent strikeforce img ID in .txt instead of modding config
The fetch_and_alert.yml file that runs with fetch_and_alert.py
contained the last id number of the most recent image processed
from strikeforce, and this number would be modified to reflect the
most recent id with each run. This resulted in a config file that
was being constantly modified.
Now, get_images.py that is called with fetch_image_api within
fetch_and_alert.py looks for an existing .txt file, creates one
if it does not, and reads in the id number (currently
fills it with 0 if there was no file) and updates it after
the last image is fetched for each run. This also means that
get_images.py no longer needs to read the entire .yml file
with each run, and can use the stored config values from when
fetch_and_alert.py read it in.