Add docstrings and ability to handle ER implementation
I added docstrings to both fetch_and_alert dependencies get_images
and detect_img to help explain the code and how it works. I also
added a pathway within detect_img to handle posting to Earthranger
instead of sending email alerts, but the necessary functions for that
will be added in a later commit apart of a different branch. As of this
moment, it will only send email alerts but must check the use_variation
value from config.