Add a way to attach classified images to Earthranger instance
This script defines a function that allows a classified image to be
attahced to an earthranger event, linked to the specific camera the
image came from. It gets called in detect_img if the config file
tells it to send to earthranger and not smtp emails, and if an event
was created for that specific image to be posted to. It returns
the http request response code for validation of the post.