Add functionality to create report on ER map
Functions cam_location and post_event were added, post_event calls
cam_location inside it so that it can post the event to the
correct geogaphical location on the Earthranger instance that the
image to be posted came from. The camera name as it is in ER and
auth tokens from config are all that are needed and it spits out the
unique id for the new event so that the image can subsequently
be posted to that event.