Post historical cougar sightings at each camera site in ER
The script post_cougar_log.py creates a function called is_cougar
that gets called in detect_img.py only if the label of the sorted
animal is a cougar. It then adds an observation noting the time of
upload (not the actual time the image was taken, if you want that
you'd have to click to see the actual image in the event that was
created, luckily, with the time on the historyical data and because
one would know the camera location, it would be easy to see the
image to check the actual time the photo was taken). This way
there is now a log of cougar sightings per camera.

It would be relatively simple to make this historical log capture
more than just the cougar if there are other animals we would like
this log for.