Add logging capability to fetch_and_alert
Upon running, a fetch_and_alert.log file will be created
and all debug messages get appended to that same file with each
run, regardless of if the program was stopped and restarted. It
also prints all warning and critical logging messages to the
console. Currently, no manually added logging has been added to
the script but with each run it automatically creates debug
messages specifically pertaining to the http requests.