Make cougarvision handle 2 stikeforce account pulls
We just had cougarvision running for the bioreserve.cam account,
which would get images and send alerts for that singlular account.
A few changes had to be made in order to add the overnight team
cameras, and there are still improvements to be made if one were
to add a third. But for now, this works with two strikeforce
accounts, and the config .yml also has to be adjusted for this to
work.