Fix env.yml creation and .yml config file
There were errors that occurred when creating the environment
with the current cougarvision_env.yml file, and the name of
the environment was cougarvisionTest and not cougarvision. Now the
file will allow cougarvision to run immediately upon env creation
without updating dependencies. The fetch_and_alert.yml config file
was also updated with more descriptors for each item