Add logging and error exception for camera ID's not in config
For demo purposes as well as robustness of the software, cougarvision
needs to be able to handle circumstances when the config file
camera IDs don't contain all the cameras that may be sending photos to
strikeforce. Now, get_images will log a warning in the
cougarvision.log file that alerts the user if there was a photo
retrived from a camera it doesn't regognize so that ID can
be added, without it crashing the whole program and allowing it
to continue to run. It also doesn't muddy up the terminal when
running a demo, as most camera images are skipped in that
circumstance

Resolves #41