Integrated most updated animl-py master branch
Some changes were made to animl-py and so fetch_and_alert.py
and detect_img.py had to be changed to account for function name
differences. strikeforcegetcameras.py was also added so that users
can obtain their camera ids with one script rather than looking
for them in the web dev of strikeforce.