resolve pip install errors from animl update
The newest version of animl on pypi that is pip installed removes
the detector initialization line, and is instead initialized when
the megadetector class is initialized. Also, there were some
function changes to the flow of detection, wherein the generator
internally adds the boxes without a line within cougarvision
detect_img.py.

Resolves #53