Make the demo modular
The demo branch was created hard and fast and these commits
make the demo portion more modular in the sense that with
merely a few extra lines to the config, you can make fetch_and_alert
run the necessary portions to allow the output to be visualized.
Currently, one must run this version of fetch_and_alert and
simultaneously run display.py with the same config
file as fetch and alert to work. display.py is the main function
and visualize_helper.py contain some functions that get called
in fetch_and_alert to ensure the images are in the proper place
and format for display.py to reach them.