Fix logging duplicates
The script was not writing the log to a file, only to the terminal
and was logging duplicated lines. It needed to be initialized
before calling the classifier and detector initializers because
when those got called first, they introduced their own logging
handlers that overrode the one I created. Also the force=True
needed to be added so that the log file would actually be created