Handle .pt classifier models
We trained a new model 'southwest_v3.pt' that performs better
on cougars than the previous model, but the animl-py on pypi
can only handle h5 models. Cougarvision had to be updated to be
compatible with the updates that have been made to animl-py
on github where there exists the capability to use .pt models. It
was mostly function name changes and movement, and the updates
only affected fetch_and_alert.py for loading the classifier model
and detect_img.py.